+++ src/main/java/de/uni_stuttgart/it_rex/course/service/dto/written/CourseDTO.java
@@ -1,0 +1,4 @@
+package de.uni_stuttgart.it_rex.course.service.dto.written;
+
+public class CourseDTO {
+}
