+++ src/main/java/de/uni_stuttgart/it_rex/course/service/dto/written/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom DTOs.
+ */
+package de.uni_stuttgart.it_rex.course.service.dto.written;
