+++ src/main/java/de/uni_stuttgart/it_rex/course/service/validation/written/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Custom service layer beans for validation.
+ */
+package de.uni_stuttgart.it_rex.course.service.validation.written;
