+++ src/main/java/de/uni_stuttgart/it_rex/course/web/rest/written/ChapterResource.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.web.rest.written;
 
-import de.uni_stuttgart.it_rex.course.domain.Chapter;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.Chapter;
 import de.uni_stuttgart.it_rex.course.service.written.ChapterService;
 import de.uni_stuttgart.it_rex.course.web.rest.errors.BadRequestAlertException;
 import io.github.jhipster.web.util.HeaderUtil;
@@ -178,5 +178,4 @@ public class ChapterResource {
             true, ENTITY_NAME, result.getId().toString()))
         .body(result);
   }
-
 }
