+++ src/test/java/de/uni_stuttgart/it_rex/course/domain/written/ChapterTest.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.domain.written;
 
-import de.uni_stuttgart.it_rex.course.domain.Chapter;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.Chapter;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
