+++ src/test/java/de/uni_stuttgart/it_rex/course/domain/written/CourseTest.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.domain.written;
 
-import de.uni_stuttgart.it_rex.course.domain.Course;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.Course;
 import de.uni_stuttgart.it_rex.course.domain.enumeration.PUBLISHSTATE;
 import org.junit.jupiter.api.Test;
