+++ src/test/java/de/uni_stuttgart/it_rex/course/utils/written/ContentIndexUtil.java
@@ -1,7 +1,6 @@
 package de.uni_stuttgart.it_rex.course.utils.written;
 
-import de.uni_stuttgart.it_rex.course.domain.ChapterIndex;
-import de.uni_stuttgart.it_rex.course.domain.ContentIndex;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.ContentIndex;
 
 import java.util.List;
 import java.util.UUID;
