+++ src/test/java/de/uni_stuttgart/it_rex/course/web/rest/mapper/written/CourseMapperTest.java
@@ -1,7 +1,7 @@
 package de.uni_stuttgart.it_rex.course.web.rest.mapper.written;
 
 import de.uni_stuttgart.it_rex.course.domain.enumeration.PUBLISHSTATE;
-import de.uni_stuttgart.it_rex.course.domain.Course;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.Course;
 import de.uni_stuttgart.it_rex.course.service.mapper.written.CourseMapper;
 import de.uni_stuttgart.it_rex.course.service.mapper.written.CourseMapperImpl;
 import org.junit.jupiter.api.Test;
