+++ src/test/java/de/uni_stuttgart/it_rex/course/web/rest/written/ChapterResourceIT.java
@@ -2,8 +2,8 @@ package de.uni_stuttgart.it_rex.course.web.rest.written;
 
 import de.uni_stuttgart.it_rex.course.CourseServiceApp;
 import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
-import de.uni_stuttgart.it_rex.course.domain.Chapter;
-import de.uni_stuttgart.it_rex.course.domain.ContentIndex;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.Chapter;
+import de.uni_stuttgart.it_rex.course.domain.written_entities.ContentIndex;
 import de.uni_stuttgart.it_rex.course.repository.written.ChapterRepository;
 import de.uni_stuttgart.it_rex.course.repository.written.ContentIndexRepository;
 import de.uni_stuttgart.it_rex.course.web.rest.TestUtil;
